Update RestVerifyModel.cs in 4.60.6.1 version.

+++ 4.60.6.1/NopTop.Plugin.Payments.Zarinpal/Models/RestVerifyModel.cs
@@ -17,7 +17,7 @@ public class RestVerifyData
     public string message { get; set; }
     public string card_hash { get; set; }
     public string card_pan { get; set; }
-    public int ref_id { get; set; }
+    public long ref_id { get; set; }
     public string fee_type { get; set; }
     public int fee { get; set; }
 }
